add CloseTarget to close targets
Bit hacky -- creates a client and calls CloseTarget on that client. We
can get the list of targets with ListTargets and close any of those.
Since we only identify targets by ID outside the client, we have to
create a dummy target to pass to the client.

Need to figure out if we leak too many TargetHandlers when running this
for a while.